Automate Releases Using Travis CI and Github Releases
Add Travis CI config with github release support.
This should upload artifacts to GitHub release after
a tag is pushed to the repo. A way to automate the release.
It also makes builds on every branch that gets a commit push.

Add information on how to make a release for the project. This
is a separate markdown file detailing the procedure.