FTDISerialDevice: simplify adaptArray() length formula and fix crash bug
The old iterative method of calculating realLength was pretty clumsy
and slow.

Additionally, it failed to consider an incomplete FTDI header.  This
shouldn't happen, but it was seen in the dirty real world:

 java.lang.ArrayIndexOutOfBoundsException: src.length=65 srcPos=2 dst.length=61 dstPos=0 length=62
   at java.lang.System.arraycopy(Native Method)
   at com.felhr.usbserial.FTDISerialDevice.copyData(FTDISerialDevice.java:527)
   at com.felhr.usbserial.FTDISerialDevice.adaptArray(FTDISerialDevice.java:508)
   at com.felhr.usbserial.UsbSerialDevice$WorkerThread.doRun(UsbSerialDevice.java:359)
   at com.felhr.usbserial.AbstractWorkerThread.run(AbstractWorkerThread.java:21)

The same bug would manifest in copyData() due to a badly implemented
length check.

Fixes https://github.com/XCSoar/XCSoar/issues/1032